Install WSL if necessary
Outdated digression follows:

It seems that wsl --status returns 0 when WSL is available and
something else (50 in my case) when it is not.

The output of wsl --help also differs, with the uninstalled case
printing the output of a shim that pretty much can only install
WSL, and an optional distribution in one go.

WSL can be installed using the "optional components" system or from
the Windows store. Here we choose the version from the optional
components.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>