Fix WSL installation, add caching, cloud-init
The image is pulled from cloud-images, initialized with cloud-init. A
test user is added so that we can see (later, not yet) snaps running as
a non-root user.

WSL uses UTF-16 by default, which is not compatible with the output
encoding of actions runner. Use WSL_UTF8=1 environment variable to opt
into the more universal encoding.

Silence download progress (Invoke-WebRequest) for huge speedup by
setting $ProgressPreference = 'SilentlyContinue'

Install WSL from the web download, avoiding both the old version in the
base system and the up-to-date version in the store that is hard to
install without an active desktop session.

Disable verbose tracing, add more error checks and print outputs to
better explain what is going on.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>