Fix error handling in "install $SNAP" step
WSL error handling involving native commands is somewhat tricky, so
after using Start-Process I'm not a lot more inclined to revert back to
the simples direct dispatch.

The sequence used tests out okay in practice. A spurious print of a
false value may be printed in case of failure, but this is harmless.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>